+ ressource_component, change ressource_view
